refactor: Replace collect by home
